+++ flowc/src/lib/dumper/mod.rs
@@ -1,4 +1,6 @@
 /// Module to dump a flow, or functions to .dot files of directed graphs
-pub mod dump_dot;
+pub mod flow_to_dot;
 /// Module to dump a flow, functions or tables to text files for inspection
 pub mod dump;
+/// Module to output the graph of functions after compilation
+pub mod functions_to_dot;
